+++ NEWS.md
@@ -1,7 +1,11 @@
+# pedquant 0.2.2
+
+* replace stock data source from the 163 to the Eastmoney
+* fixed mutual bugs
+
 # pedquant 0.2.1
 
 * remove several not needed scripts
-* replace stock data source from 163 to eastmoney
 
 # pedquant 0.2.0